Fixed undefined behaviour. Added badges

+++ src/OpenSimplex2F.c
@@ -846,7 +846,7 @@ OpenSimplexGradients *newOpenSimplexGradients(OpenSimplexEnv *ose, int64_t seed)
         source[i] = i;
     }
     for (int i = PSIZE - 1; i >= 0; i--){
-        seed = seed * 6364136223846793005L + 1442695040888963407L;
+        seed = (uint64_t)seed * 6364136223846793005ULL + 1442695040888963407ULL;
       int32_t r = (int32_t)((seed + 31) % (i + 1));
 		if (r < 0){
             r += (i + 1);
